requirements.txt
anyio==3.5.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asttokens==2.2.1
Babel==2.11.0
backcall==0.2.0
bcrypt==3.2.0
Bottleneck==1.3.5
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.0.4
colour==0.1.5
comm==0.1.2
contextvars==2.4
contourpy==1.0.5
cryptography==39.0.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fonttools==4.25.0
furl==2.1.2
greenlet==2.0.1
h11==0.12.0
htmltools==0.1.4
idna==3.4
immutables==0.19
infinity==1.5
intervals==0.9.2
ipykernel==6.21.3
ipython==8.11.0
ipython-genutils==0.2.0
ipython-sql==0.3.9
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
linkify-it-py==2.0.0
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.0
munkres==1.1.4
nest-asyncio==1.5.6
numexpr==2.8.4
numpy==1.24.3
orderedmultidict==1.0.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
passlib==1.7.4
pendulum==2.1.2
pexpect==4.8.0
phonenumbers==8.12.27
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==3.1.0
prettytable==3.5.0
prompt-toolkit==3.0.38
psutil==5.9.0
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.6
pytz==2022.7
pytzdata==2020.1
pyzmq==25.0.0
requests==2.29.0
rsconnect-python==1.16.0
seaborn==0.12.2
semver==2.13.0
setuptools==66.0.0
shiny==0.2.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==2.0.12
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.3
stack-data==0.6.2
starlette==0.26.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
uc-micro-py==1.0.1
urllib3==1.26.15
uvicorn==0.20.0
wcwidth==0.2.6
websockets==10.4
wheel==0.38.4
